added spec to ensure secure connection
