+++ spec/ftp_spec.rb
@@ -42,6 +42,38 @@ describe CarrierWave::Storage::FTP do
     @stored = @storage.store!(@file)
   end
 
+  it "opens/closes a secure ftp connection to the given host" do
+    CarrierWave.configure do |config|
+      config.reset_config
+      config.ftp_host = 'ftp.testcarrierwave.dev'
+      config.ftp_user = 'test_user'
+      config.ftp_passwd = 'test_passwd'
+      config.ftp_folder = '~/public_html'
+      config.ftp_url = 'http://testcarrierwave.dev'
+      config.ftp_passive = true
+      config.ftp_tls = true
+    end
+    @secure_storage = CarrierWave::Storage::FTP.new(FtpUploader)
+
+    ftp = double(:ftp_connection)
+    ftp_params = [
+      'ftp.testcarrierwave.dev',
+      'test_user',
+      'test_passwd',
+      21
+    ]
+
+    Net::FTP.should_receive(:new).and_return(ftp)
+    ftp.should_receive(:connect).with('ftp.testcarrierwave.dev', 21)
+    ftp.should_receive(:login).with('test_user', 'test_passwd')
+    ftp.should_receive(:passive=).with(true)
+    ftp.should_receive(:mkdir_p).with('~/public_html/uploads')
+    ftp.should_receive(:chdir).with('~/public_html/uploads')
+    ftp.should_receive(:put).with(@file.path, 'test.jpg')
+    ftp.should_receive(:quit)
+    @stored = @storage.store!(@file)
+  end
+
   describe 'after upload' do
     before do
       ftp = double(:ftp_connection)
