+++ spec/spec_helper.rb
@@ -2,6 +2,7 @@ require 'rubygems'
 require 'bundler/setup'
 require 'rspec'
 require 'tempfile'
+require 'pry'
 
 require 'carrierwave'
